Further restrict mount interface slots
With the upcoming port-forward interface, we need to expand the workshop
definition docs, as this is the only place where attributes like host-to
and host-from make sense to explain. These docs are much easier to read
when the interfaces behave consistently.

The system SDK will now only have one mount slot, named "mount". This is
the same as most other interfaces, and reinforces the idea that the
system SDK is for system resources only (no workshop-source on a system
SDK slot).

There's no need to prevent installation of slots with a host-source
attribute at the policy level, as this is handled by the slot sanitizer,
in line with other interfaces. This attribute is never attached to the
slot itself, only to a ConnectedSlot.

I think the previous auto-connect policy would allow auto-connecting any
slot named "mount," which doesn't make sense to me because this slot
might not have the special host-source logic that the system SDK does.
So auto-connect is now based on the slot SDK type or explicit
connections from the workshop definition.