Add interfaces to workshop definition reference
Most interfaces behave differently in the system SDK, so I think they
should be documented alongside the workshop definition rather than the
SDK definition.

There will probably be a fair bit of repetition between these two
references, but I think the table-based format is a nice alternative to
the example-based format in the SDK definition docs, and it doesn't hurt
to have both. If it gets out of hand we could document the
system-specific attributes in the workshop definition, and refer to the
SDK definition for other attributes.

This should also make it easier to remember to update the workshop
definition docs when adding a new interface.

Co-authored-by: Artem Konev <141050460+akcano@users.noreply.github.com>
Co-authored-by: Dmitry Lyfar <69887876+dmitry-lyfar@users.noreply.github.com>