LaunchWorkshop to support rebuilding a workshop if existed
Change the LaunchWorkshop name to
LaunchOrRebuildWorkshop to support a scenario when
a workshop exists and it needs to be 'factory
reset' using a provided base image.

This change allows refresh to not remove the
existing workshop if a full rebuild is required
due to the base change.

Thus, any refresh plan will reuse the existing
workshop by rebuilding or restoring it from a
provided snapshot. That would enable minimal
interface changes required for the backend
implementation. Also, it simplifies logic in
workshopstate/request.go as both `launch` and
`refresh` would reuse bits of the launch workshop
logic.