Swap prepareRefresh and workshop conflict check in RefreshMany
* RefreshMany should unlock the state for the
duration of the store requests. This allows for
other requests to be processed. The store requests
are now resolved before any conflict check is made
for the workshop(s). It allows to unlock the state
for the duration of the requests and disallows
concurrent refreshes of the same workshop at the
expense of doing extra work and abandoning it if
there is a conflicting refresh change that's been
spawned in the meantime.

* Some naming updates.