Install xdelta3 to workaround download issue with snapd deb
If certain store requests are rate-limited, snapd fails to retry unless
xdelta3 is installed. This is not a dependency of the Debian package,
but is included in the snap.

Since we need to install xdelta3 for all 3 test suites, I opted to
install the other deb packages (and yq) for the integration tests, to
keep things simple (even though they aren't needed).

Co-authored-by: Dmitry Lyfar <69887876+dmitry-lyfar@users.noreply.github.com>