Treat system SDK more like a Store SDK
In order to bring the system SDK in line with other SDKs, and to allow
future updates to this SDK without removing existing workshops, we now
track revision numbers for the system SDK. At the moment these are
hard-coded, protected by a test (to remind us to update it as needed).

The system SDK could be mounted from a local directory or a LXD volume.
Since only one copy is needed for all users, a volume makes more sense.