Run setup-base hook and snapshot right after installing SDK
Previously, Workshop made an effort to remove existing SDKs after
restoring from a snapshot. However, the removed SDKs are based on the
current refresh. Future refreshes won't know about these SDKs, but
volumes are still mounted in the snapshot.

This commit restores earlier behaviour, where setup-base is run after
installing the associated SDK, and before other SDKs are mounted.

Other SDKs no longer need to be removed from the workshop after
restoring a snapshot. This means `FakeWorkshopBackend.Restore` needs to
be a bit smarter at restoring snapshots. Rather than copying the
WorkshopFs on each snapshot, it's enough for now to just remove the SDKs
that were installed after the current snapshot.

Rebuilding a workshop will now remove all SDK mounts, to be consistent
with the new restore behaviour (and launch). This requires a workaround
for the project mount, which in a future bugfix will only be mounted
after all snapshots are created.