Rework default config and devices for rebuilding workshops
The default config and devices are merged into the existing ones. This
is the correct behaviour for config, because LXD sets various options
which are outside our control. However, the default config now includes
a few more options to ensure we reset the ones we do control.

Removing all existing devices will allow us to skip unmounting SDKs
after rebuilding. To avoid removing the apt cache mount here, it has
been added to the list of default devices. Currently the project mount
is available during setup-base hooks, which is a bug because workshop
snapshots should only depend on the base and list of SDKs. Moving the
mount task will be a breaking change which needs to carefully managed,
so for now this mount will be manually copied over when rebuilding.
Eventually rebuild should have the same observable effects as remove +
launch, ignoring caches and possibly some other details.

An earlier attempt did move the project mount to the list of default
devices. This uncovered a bug in the backendDeviceSuite, where workshops
were being created with random project IDs and the apt cache volumes
weren't being cleaned up. This has been fixed by mocking `NewProjectId`
and removing the test workshop, in line with other integration tests.