Retrieve base images in parallel with SDKs
Previously, if the download-base task failed during a refresh, the
stash-workshop task could not be undone, because the unstashed workshop
still existed. The create-workshop task has to immediately follow
stash-workshop, and remove the unstashed workshop if it fails.

It's still possible for create-workshop to leave the workshop intact if
it fails due to an internal error. There isn't an easy way around this
because the project ID needs to be known in order to remove the
workshop, and retrieving the ID from the task can fail. So for now I've
left this part as-is. It's unlikely that users will encounter internal
errors anyway.

Since download-base needs to happen earlier, it might as well be
concurrent with the SDK downloads.