Use host directory for apt cache
The workshop data can be populated in various ways during launch (e.g.
from mount plugs, local SDKs and the apt cache). This data should be
removed if launch fails, and when the workshop is removed, but not when
refresh fails. To this end we expand the apt-cache tasks to cover all
host directories that Workshop manages.