Preserve DHCP lease during refresh
LXD will send a DHCP release when removing an instance with a known
physical address. As a workaround we can remove the address before
deleting copies of the workshop or stash.

When a workshop is removed manually, or during a failed launch, the DHCP
lease will be cleared as usual.