Prevent DHCP release when workshop stops
When a workshop is stopped manually, systemd-networkd sends a DHCP
release request to the bridge. This does not seem to occur when a
workshop is stopped during refresh. In any case, the request is ignored
by dnsmasq because it doesn't contain a server ID field. Since this may
change in future, it's safer to disable the request entirely.

The current fix relies on a file name generated by netplan. We should
probably use netplan directly if/when it supports SendRelease.