Rework LxdProjectName to accomodate non-alphanum usernames
If a user name is non-POSIX, it could bring
issues, e.g. having an '@' in the user name breaks
snapshot logic as snapshots use project name and
preceive '@' as a separator. To avoid any further
conflicts, we only allow an alphanum user name,
otherwise, a UID for the user will be used for
the LXD project name.