Prevent multiple desktop and ssh-agent connections
Currently the ssh-agent interface creates one script in /etc/profile.d
per plug. This doesn't really make sense, and if we add attributes to
the ssh-agent it may result in an inconsistent environment. The desktop
interface shares this script between plugs, and is removed when any of
them are disconnected.

To fix, we just prevent connecting more than one ssh-agent or desktop
interface plug. If users really want to connect two or more, they can
use plug binding.