Restrict unix socket paths inside the workshop
LXD 6.3 and 6.4 clean up socket files on the host instead of the
instance, so it's not safe to create a Wayland socket inside the
workshop - the real one is deleted on disconnect. The issue should be
fixed in the next release of LXD [1].

Luckily there's a separate issue which makes the first one a bit less
impactful [2]. LXD can only access the host filesystem via various bind
mounts (e.g. /home, /run/user/1000 and /var/lib/snapd/hostfs), so it's
reasonably safe to create sockets in some paths, including /tmp and
/var/lib/workshop/run. Luckily X11 uses the former and ssh-agent uses
the latter. For Wayland, we move the socket to /tmp and create a symlink
to it in /run/user/1000. Another option would be to make WAYLAND_SOCKET
an absolute path, but I'm not sure how many clients support that
convention.

[1] https://github.com/canonical/lxd/issues/15782
[2] https://github.com/canonical/lxd/issues/15784