Restrict unix socket paths inside the workshop
LXD 6.3 and 6.4 clean up socket files on the host instead of the
instance, so it's not safe to create the default Wayland socket inside
the workshop - the real one is deleted on disconnect. The issue should
be fixed in the next release of LXD [1].

Luckily there's a separate issue which makes the first one a bit less
impactful [2]. LXD can only access the host filesystem via various bind
mounts (e.g. /home, /run/user/1000 and /var/lib/snapd/hostfs), so it's
reasonably safe to create sockets in some paths, including /tmp and
/var/lib/workshop/run. Luckily X11 uses the former and ssh-agent uses
the latter.

Unfortunately, snapd requires the Wayland socket to match
/run/user/1000/wayland-[0-9]*. This is a glob, not a regex, so we can
append a string unique to workshop to ensure that the host's socket
remains intact.

[1] https://github.com/canonical/lxd/issues/15782
[2] https://github.com/canonical/lxd/issues/15784