Attempt to stop gracefully before forcing stop
I noticed stopping sometimes takes an abnormally long time during stash.
This is because stash calls stop with force=true. Although force
stopping uses SIGKILL, which is fast, unmounting the container
filesystem can still take around 5 seconds, which is slower than a
typical graceful shutdown. I'm not sure why unmounting is slow, but it
might be related to ZFS running maintenance tasks in the background.

Given this, it makes sense to shutdown gracefully instead of forcibly.
This also gives us access to shutdown logs in case we want to optimise
shutdown further in future. If the graceful shutdown fails or times out,
the workshop could be running, stopped or somewhere in between, so we
proceed to kill it just to be sure. This matches what LXD itself does
when evacuating a cluster.