Add fsutil package
The new sftpFs works around a number of issues with afero/sftpfs:
- Mkdir and OpenFile aren't transactional, they ignore the umask,
  and don't return ErrExist (required for afero.TempDir and
  afero.TempFile).
- Paths are stripped from certain error types
- On error, functions can return non-nil file handles
- Sync and WriteAt aren't implemented at all
- Readdir and Readdirnames don't use the open file handle

The Fs wrapper also brings MkdirAll, MkdirTemp, CreateTemp, WriteFile
and ReadFile in line with the standard library implementations.

It turns out the only functions that sftpfs implements correctly are
those which directly call the corresponding function on the client, so
there's no need for sftpfs any more. But we'll continue using
BasePathFs. All of these have moved to a standalone fsutil package.

Co-authored-by: Dmitry Lyfar <69887876+dmitry-lyfar@users.noreply.github.com>