Handle attribute errors when connecting mount interface
I was ignoring errors based on the assumption that plugs and slots are
always sanitised before connecting them. This is true in all cases I'm
aware of, but it's possible for the attributes to change after they've
been sanitised, since the state is encoded using JSON. I've relaxed the
conversion logic to allow reasonable float -> int conversions, but I
don't think it's safe to rely on the attributes anymore.