Prevent concurrent TICS server jobs
Previously, pushing to main would cancel the previous run. If the cancel
happens in the middle of the TICS step, the next run may fail because
TICS has yet to be cancelled properly. Now we'll only cancel jobs for
pull requests, which run TICS in client mode.

I also removed the fallback to `github.run_id`. All the events we
support define GITHUB_REF [1]. The linked stackoverflow post is based on
the official docs [2], which only use a fallback for more obscure
settings like `github.head_ref`.

[1] https://docs.github.com/en/actions/reference/workflows-and-actions/events-that-trigger-workflows
[2] https://docs.github.com/en/actions/how-tos/write-workflows/choose-when-workflows-run/control-workflow-concurrency