Wait for systemd to be ready before waiting for systemd-networkd
This test is flaky on my "update base image" branch, which is odd
because I've never noticed it being flaky before, and I don't think I
made any major changes to the code this exercises. In any case, it
should be more reliable now.

The error manifests as "command exit code 1" when LaunchTestWorkshop
runs the start command. The stderr of the start command is "Could not
create manager: No such file or directory." After installing strace in a
custom base image, and looking at the code a bit, I was able to confirm
that the file in question is /run/systemd/. So this particular error is
definitely fixed by waiting for systemd to be ready.

I probably should have done this from the beginning, the extra Exec call
is much simpler than I was imagining.