Separate GetBase from DownloadBase
Adds a new method GetBase to BaseImageManager. It looks up the latest
version of a given base. For the LXD backend, the version is a
fingerprint. This will allow a future change to make the base image part
of the refresh plan.

Also renames Download to DownloadBase. It now accepts a version instead
of returning one.

I've changed my mind about downloading using the alias instead of the
fingerprint. It should be reasonably safe to use fingerprints, since the
image servers I know of keep their images around for at least a day
after updating them. Also, allowing DownloadBase to change the
fingerprint feels like leaking LXD quirks into the generic API.

The downside of using fingerprints is that local images won't have an
UpdateSource. LXD doesn't use this because we disable auto-updates.
Workshop also won't need it if we use a custom property instead, similar
to the existing user.workshop.* instance config options.

The new property is "workshop-base." I avoided "user.workshop.*" so we
can filter by properties.workshop-base (dots in filters are treated like
path separators). I don't think there's a high chance of name collisions
here.

In order to apply the property, I had to use CreateImage instead of
CopyImage. The behaviour should be identical for simplestreams servers,
but for LXD image servers with multiple addresses, only the main address
will be used. As far as I can tell, servers only have multiple addresses
when they listen on multiple network interfaces (on the same machine),
so I don't see why you would want to use them over the main address. In
any case, we can implement this functionality later if we need to.