Avoid removing workshop when restoring stash
LXD supports "incremental" or "refresh" copies, which means it replaces
the instance's root storage volume without ever removing the instance
itself. When doing this, it does not update the instance's config or
devices for some reason (the same happens when rebuilding from a base
image).

Currently, when restoring the stash, we first clear the workshop's MAC
address so that LXD doesn't try to clear its DHCP lease. This is no
longer necessary, since we now copy the stash over the existing
workshop.

However, removing the stash could potentially lead to the same issue. To
prevent this, we omit the MAC address and other volatile attributes when
creating the stash. This is handled by LXD automatically, by passing an
empty config map to the copy operation. When restoring the stash, we
manually copy over the same attributes: all the non-volatile ones plus
the base image and idmap. The corresponding function in LXD is called
InstanceIncludeWhenCopying.