Limit workshop and SDK name lengths
SDK names are already limited to 40 characters by SDKcraft. We should
use the same limit in Workshop so that SDK layers don't exceed the LXD
instance name limit of 63 characters.

Workshops are already used in instance names. In theory the max length
is 63 - 9 = 54 but we might as well use the same limit as SDKs.