Use snap hooks to install completion scripts for bash, fish and zsh
Snap's completion support is geared towards strictly confined snaps and
only supports bash. It might be possible to adapt it to zsh by writing a
zsh-style completion script which calls the bash-style one inside the
snap. But I doubt it can be done in a robust way. Reimplementing the
bash approach for both fish and zsh looks to be quite difficult.

Instead, since Workshop uses classic confinement, we can just install
completion scripts on the host. Zsh looks for scripts in $fpath, which
by default only has one entry outside /usr/share. However, bash and fish
both look for completion scripts in $XDG_DATA_DIRS, so we can use
/var/lib/snapd/desktop for them. This is where snapd itself installs its
bash completion scripts, so it should be OK for us to use as well.