Query and compute hashes for Store SDKs
The current Store has a couple of issues:
- Only supports MD5 hashes
- Only stores one revision at a time

For now, we'll use MD5 for Store SDKs but continue using SHA3 for other
SDKs, in line with the Snap Store. Since the SDK we download may have
been updated after the initial query, we also compute the hash during
the download. We still need *a* hash to plan a launch or refresh, but
once we've used it to decide to download a new SDK the original value
can be discarded.