Implement a git-like directory hash
The goal is to hash sketch and in-project SDK contents, for caching
purposes. We can also use it to compare the source files (e.g. from the
project directory) to the committed SDK revisions.

Since we ignore unusual files (anything that's not a regular file,
directory or symlink), and only preserve data + permissions, we reuse
git's metadata format, which consists of a "blob <length>\0" prefix for
files and symlinks, and a sequence of "<mode> <filename>\0<hash>",
prefixed by "tree <length>\0", for directories. We only match git in
simple cases, because it uses SHA1 (which we'd rather avoid) and only
supports 644 and 755 permissions.