Introduce BaseImage struct
This is intended as an equivalent of sdk.Setup for base images. It
introduces duplication of the base's name in some places, but I think
the code is ultimately more readable.

The hash for an SDK layer will be determined by a BaseImage together
with an []sdk.Setup. These will likely end up as fields in a struct,
so it's nicer to bundle the image-related data into a single field.