Add hash to workshop snapshots
A hash is constructed using a similar format to osutil.HashDirEntries,
which is based on git tree hashes. The initial data consists of a fixed
number, representing the daemon version, followed by the base name and
fingerprint. Each SDK contributes another chunk of data: the SDK kind
(volume or local), name and hash.

We don't poison the cache with every new daemon version; instead a
hard-coded revision number should be bumped every time we change the
format of workshop snapshots. This could be a change to the workshop's
filesystem, its config options or devices, among other things.

I've made my best attempt at ensuring we keep this up to date, by adding
integration tests for the LXD backend and the daemon API. The tests
generate some data and compare it against a known good reference. In
theory we could update these reference files automatically.