Rename ImportVolume to ImportSdk in backend
The VolumeSetup and sdk.Meta structs have converged to the point where
it doesn't make sense to separate them once the volume-related functions
are split into SDK and state specific ones. For now I patched up a few
places which still use it, but removed it from the API itself.