Adjust revision timestamps in TestCommitKeepsInstalled
In bb0d2c82ce21b9c9449c811652666f89daf7f2ea I introduced hashes for
local SDKs. To track the oldest SDK, timestamps are stored in revision
symlinks which point to the hash (a directory containing the SDK). For
some reason I neglected to migrate from `os.Chtimes` to `sys.Lchtimes`
in this test. This results in an extremely rare error when revision 44
is deleted instead of 42, if they end up with the same timestamp.