Decouple device names from volume names
Workshop snapshots only care about the content hash of an SDK, not its
revision number. For example, a workshop might be refreshed after trying
out an SDK and uploading it to the Store. This can change the revision
number from negative to positive, without affecting the SDK at all. In
order to reuse the snapshot, it's much easier to replace an existing
disk device than to remove the old one and add a new one. This means we
should remove the revision number from the device name.

I also considered replacing the revision number with a hash, in the
volume name. This doesn't really work because the revision number is
relevant for user-facing info like the output of `sdk list`. But if we
did at some point, the volume names would be too long to fit in a device
name.