Handle ambiguous workshop exec and run arguments
Using a separator (--) to split the command or action from workshop CLI
arguments leaves no room for ambiguity. But users seem to prefer leaving
it out, for interactive use. So we now parse arguments like this:

- `workshop exec dev -- bash`: execute "bash" in "dev"
- `workshop exec -- bash`: execute "bash" in the single workshop
- `workshop exec dev bash`: try the first case, but fall back to the
  second if:
  - the project has a single workshop
  - it's not called "dev", and
  - stdout is a TTY.

The workshop CLI now ignores `--` if it's not in position 0 or 1. This
is mostly to simplify the code, but also these are the only places where
it has a special meaning. For reference, sudo only considers `--` if
it's before the command.

Co-authored-by: Artem Konev <141050460+akcano@users.noreply.github.com>