Add InstallOrder to SdkInstallation
This makes it easier to sort SDKs without having to rely on install
times or the workshop file. I plan on removing the workshop file from
snapshots, but I think it should be possible (e.g. for debugging) to
look at a snapshot and work out how it was built. For that, we need to
know which order the SDKs were installed in.