Handle EBUSY when remounting across filesystems
Some filesystems report EBUSY rather than EXDEV when Workshop attempts
to remount a plug over them. This was initially reported for 9p in a
multipass VM. I was able to reproduce in a LXD VM by disabling virtiofs
(by bind-mounting a file over /snap/lxd/current/bin/virtiofsd to remove
its execute permissions). An easier way to reproduce is to try to
remount something over /boot in a LXD VM with Ubuntu 24.04.