Split LaunchMany and RefreshMany into two phases
The first phase queries the image server, SDK Store and local SDK
storage areas, returning a "manifest" describing the ingredients that go
into the launched or refreshed workshop. For refresh, we also retrieve
the manifest of the existing workshop for comparison.

The second phase turns the manifest(s) into a set of Tasks that performs
the launch or refresh.

The reason to split it like this is that I want to attach parts of the
manifest, mainly the SDKs, to the Change rather than individual Tasks.
But I think it also makes the code a bit nicer.

I added comprehensive tests for the first phase. The second phase
already has decent coverage from the API tests, but could maybe be
improved with a few unit tests. I'll defer adding these until the
upcoming rework of the second phase to support global snapshots.